* Split VisualLog into several smaller classes to make it more maintainable
