+++ scistag/tests/filestag/__init__.py
@@ -3,11 +3,12 @@ Contains the tests for the FileStag module
 """
 from scistag.tests.visual_test_log_scistag import VisualTestLogSciStag
 
-vl = VisualTestLogSciStag(test_filename=__file__)
+test_log = VisualTestLogSciStag(test_filename=__file__)
+vl = test_log.default_builder
 
 
 def teardown_module(_):
     """
     Finalize the test
     """
-    vl.finalize()
+    test_log.finalize()
