+++ scistag/tests/logstag/__init__.py
@@ -4,12 +4,13 @@ Implements the tests for the logstag module
 
 from scistag.tests.visual_test_log_scistag import VisualTestLogSciStag
 
-vl = VisualTestLogSciStag(test_filename=__file__,
-                          formats_out={"html", "md", "txt"})
+test_log = VisualTestLogSciStag(test_filename=__file__,
+                                formats_out={"html", "md", "txt"})
+vl = test_log.default_builder
 
 
 def teardown_module(_):
     """
     Finalize the test
     """
-    vl.finalize()
+    test_log.finalize()
