+++ scistag/tests/shapestag/__init__.py
@@ -4,11 +4,12 @@ Executes the tests for the ShapeStag module
 
 from scistag.tests.visual_test_log_scistag import VisualTestLogSciStag
 
-vl = VisualTestLogSciStag(test_filename=__file__)
+test_log = VisualTestLogSciStag(test_filename=__file__)
+vl = test_log.default_builder
 
 
 def teardown_module(_):
     """
     Finalize the test
     """
-    vl.finalize()
+    test_log.finalize()
